LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigation-history-entry/entries-after-javascript-url-navigation-expected.txt



FAIL entries() after navigation to a javascript: URL assert_equals: expected 2 but got 1
